academic-research-accelerator/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-multipart==0.0.22
aiofiles==23.2.1
openai==1.30.1
langchain==0.3.25
langchain-openai==0.2.14
langchain-community==0.3.27
faiss-cpu==1.8.0
sentence-transformers==3.0.0
numpy==1.26.4
PyMuPDF==1.24.3
tiktoken==0.7.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
aiosqlite==0.20.0
feedparser==6.0.11
networkx==3.3
